Fix: Various UI fixes (#43)
* Remove redundant comment

* Break words for long description

* Fix default expiry

* Handle non discrete invoice expiry step

* Modify route protection logic

* Fix logic to determine expiry step

* Fix bug on back press when camera rejected

* Ensure icon sizes

* Remove log

* Remove log

* Add error handling for decoding bolt11

* Update Lnmessage

+++ src/lib/components/Scanner.svelte
@@ -146,10 +146,13 @@
   }
 
   onDestroy(() => {
-    qrWorker.removeEventListener('message', handleMessage)
-    qrWorker.terminate()
+    if (qrWorker) {
+      qrWorker.removeEventListener('message', handleMessage)
+      qrWorker.terminate()
+    }
+
     tickTimeout && clearTimeout(tickTimeout)
-    stream.getVideoTracks().forEach((track) => track.stop())
+    stream && stream.getVideoTracks().forEach((track) => track.stop())
   })
 </script>
